+++ CHANGELOG.md
@@ -1,8 +1,19 @@
 # Changelog
+
+## [1.0.0 - 2025 11 26]
+
+### Added
+- mk_core.py: add functions to add, edit, list, 
+ and remove items from a grocery list
+- mk_launch.py:  add function to interact with user to 
+    add, edit, list, and remove items from a grocery list
+
 ## [0.1.0] - 2025-11-19
 
 -The first version, including all code up to day 3 of the 14 day coder course.
+
 -This includes learning lists and tuples, dictionaries and sets.
+
 -Implemented changgelog and version control insturctions.
 
 ## [0.2.0]  - 2025-11-19
@@ -16,3 +27,5 @@
 ## [0.3.1] -2025 11-22
 ### Reformatted Text
 -reformatted text for CHANGELOG.md
+
+